+++ apps/etterlatte-brev-api/src/main/kotlin/no/nav/etterlatte/jobs/BehandleVedtaksbrevJobb.kt
@@ -1,0 +1,7 @@
+package no.nav.etterlatte.jobs
+
+class BehandleVedtaksbrevJobb {
+
+
+
+}
